fix: implement validation for Dictionary names and keys
This commit updates our Dictionary implementation to ensure we throw nice error messages when given invalid dictionary names or keys, such as a name for a non-existant dictionary or a key that is too long.

I've also updated the xqd header to have an enum defined for the return codes of the host-calls as it makes it simpler to understand what each return code stands for.